+++ packages/web/app/components/Appointments/LocationBookingForm/DateTimeRangeField/TimeSlotToggle.jsx
@@ -8,7 +8,7 @@ import { Colors } from '../../../../constants';
 import { TimeRangeDisplay } from '../../../DateDisplay';
 import { ConditionalTooltip, ThemedTooltip } from '../../../Tooltip';
 import { TranslatedText } from '../../../Translation/TranslatedText';
-import { CONFLICT_TOOLTIP_TITLE, TIME_SLOT_PICKER_VARIANTS } from './util';
+import { CONFLICT_TOOLTIP_TITLE, TIME_SLOT_PICKER_VARIANTS } from './constants';
 
 /**
  * @privateRemarks Specificity (0,5,0) to override styles (for all states, including :disabled and
