+++ packages/web/app/components/Appointments/LocationBookingForm/DateTimeRangeField/constants.jsx
@@ -1,0 +1,30 @@
+import React from 'react';
+
+import { TranslatedText } from '../../../Translation';
+
+export const TIME_SLOT_PICKER_VARIANTS = {
+  RANGE: 'range',
+  START: 'start',
+  END: 'end',
+};
+
+export const CONFLICT_TOOLTIP_TITLE = {
+  [TIME_SLOT_PICKER_VARIANTS.RANGE]: (
+    <TranslatedText
+      stringId="locationBooking.tooltip.unavailableTimeInRangeWarning"
+      fallback="All times must be available when booking over multiple times"
+    />
+  ),
+  [TIME_SLOT_PICKER_VARIANTS.START]: (
+    <TranslatedText
+      stringId="locationBooking.tooltip.unavailableFutureTimeWarning"
+      fallback="All future time slots must be available when booking overnight"
+    />
+  ),
+  [TIME_SLOT_PICKER_VARIANTS.END]: (
+    <TranslatedText
+      stringId="locationBooking.tooltip.unavailablePastTimeWarning"
+      fallback="All previous time slots must be available when booking overnight"
+    />
+  ),
+};
