+++ packages/web/app/components/Appointments/LocationBookingForm/DateTimeRangeField/utils.js
@@ -1,36 +1,6 @@
 import { addMinutes, differenceInMinutes, isValid, isWithinInterval, parse } from 'date-fns';
 import { isEqual } from 'lodash';
 import ms from 'ms';
-import React from 'react';
-
-import { TranslatedText } from '../../../Translation';
-
-export const TIME_SLOT_PICKER_VARIANTS = {
-  RANGE: 'range',
-  START: 'start',
-  END: 'end',
-};
-
-export const CONFLICT_TOOLTIP_TITLE = {
-  [TIME_SLOT_PICKER_VARIANTS.RANGE]: (
-    <TranslatedText
-      stringId="locationBooking.tooltip.unavailableTimeInRangeWarning"
-      fallback="All times must be available when booking over multiple times"
-    />
-  ),
-  [TIME_SLOT_PICKER_VARIANTS.START]: (
-    <TranslatedText
-      stringId="locationBooking.tooltip.unavailableFutureTimeWarning"
-      fallback="All future time slots must be available when booking overnight"
-    />
-  ),
-  [TIME_SLOT_PICKER_VARIANTS.END]: (
-    <TranslatedText
-      stringId="locationBooking.tooltip.unavailablePastTimeWarning"
-      fallback="All previous time slots must be available when booking overnight"
-    />
-  ),
-};
 
 export const appointmentToInterval = appointment => ({
   start: new Date(appointment.startTime),
